Look for translingual headers as well as embed the header.
We make the headers with ten = signs appended and prepended to make the
next parsing step easier to handle. At this point we are not really ready
to start categorizing input on the whole chunk level and the easiest way
to get the title information out is to do it this way.